+++ test/inputs/index.ts
@@ -5,4 +5,5 @@
 import './contextmenu-input.spec';
 import './key-input.spec';
 import './move-input.spec';
+import './wheel-gesture-session.spec';
 import './wheel-input.spec';